move urlProcessor to actions
