+++ src/actions/urlProcessor.test.js
@@ -1,7 +1,7 @@
 import urlProcessor from './urlProcessor';
-import { voteEdited } from '../../../actions/voting';
+import { voteEdited } from './voting';
 
-jest.mock('../../../utils/api/lsk/account', () => ({
+jest.mock('../utils/api/lsk/account', () => ({
   getAccount: jest.fn().mockImplementation(data => Promise.resolve({ address: '12L', username: data.username })),
 }));
